feat: add createIcon factory to decouple Icon from a single icon set
Extract the web and mobile Icon rendering into a `createIcon<Name>` factory
that binds a glyph map, font family, and typed name union. The default
`Icon` becomes `createIcon<IconName>({ glyphMap, fontFamily })`, so behavior
and the public API are unchanged, while consumers with their own icon
package can create a typed icon component that reuses all CDS rendering,
accessibility, and theming logic.

Font family is now injectable (moved off the static Linaria block on web to
an inline style) rather than hardcoded, enabling alternate icon sets without
forking the component.
