+++ packages/web/src/icons/index.ts
@@ -1,4 +1,6 @@
 export * from './Icon';
+export { createIcon, DEFAULT_ICON_FONT_FAMILY } from './createIcon';
+export type { CreateIconConfig, GlyphMap } from './createIcon';
 export * from './LogoMark';
 export * from './LogoWordmark';
 export * from './SubBrandLogoMark';
